[refactor] Split cmd/init_test.go into 6 topic files (#149)
