+++ cmd/init_test.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"encoding/json"
 	"errors"
 	"os"
 	"path/filepath"
@@ -9,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/lugassawan/rimba/internal/config"
-	"github.com/spf13/cobra"
 )
 
 func TestInitSuccess(t *testing.T) {
@@ -142,174 +140,6 @@ func TestInitExistingDirConfig(t *testing.T) {
 	out := buf.String()
 	if !strings.Contains(out, "already exists") {
 		t.Errorf("output should contain 'already exists', got:\n%s", out)
-	}
-}
-
-type agentFilesWant struct {
-	output   []string
-	noOutput []string
-	files    []string
-	noFiles  []string
-	marker   bool
-}
-
-func assertAgentFilesResult(t *testing.T, repoDir, out string, want agentFilesWant) {
-	t.Helper()
-	for _, s := range want.output {
-		if !strings.Contains(out, s) {
-			t.Errorf("output missing %q", s)
-		}
-	}
-	for _, s := range want.noOutput {
-		if strings.Contains(out, s) {
-			t.Errorf("output should not contain %q", s)
-		}
-	}
-	for _, f := range want.files {
-		if _, err := os.Stat(filepath.Join(repoDir, f)); os.IsNotExist(err) {
-			t.Errorf("file not created: %s", f)
-		}
-	}
-	for _, f := range want.noFiles {
-		if _, err := os.Stat(filepath.Join(repoDir, f)); err == nil {
-			t.Errorf("file should not exist: %s", f)
-		}
-	}
-	if want.marker {
-		content, err := os.ReadFile(filepath.Join(repoDir, "AGENTS.md"))
-		if err != nil {
-			t.Fatalf("read AGENTS.md: %v", err)
-		}
-		if strings.Count(string(content), "<!-- BEGIN RIMBA -->") != 1 {
-			t.Error("AGENTS.md should have exactly one BEGIN RIMBA marker after re-init")
-		}
-	}
-}
-
-func TestInitAgentFiles(t *testing.T) {
-	agentFilePaths := []string{
-		"AGENTS.md",
-		filepath.Join(".github", "copilot-instructions.md"),
-		filepath.Join(".cursor", "rules", "rimba.mdc"),
-		filepath.Join(".claude", "skills", "rimba", "SKILL.md"),
-	}
-
-	tests := []struct {
-		name        string
-		withFlag    bool
-		preExisting bool
-		runTwice    bool
-		want        agentFilesWant
-	}{
-		{
-			name:     "creates agent files",
-			withFlag: true,
-			want:     agentFilesWant{output: agentFilePaths, files: agentFilePaths},
-		},
-		{
-			name:        "installs into existing config",
-			withFlag:    true,
-			preExisting: true,
-			want:        agentFilesWant{output: []string{"already exists"}, files: []string{"AGENTS.md"}},
-		},
-		{
-			name:     "idempotent on second run",
-			withFlag: true,
-			runTwice: true,
-			want:     agentFilesWant{output: []string{"already exists"}, marker: true},
-		},
-		{
-			name: "skips agent files without flag",
-			want: agentFilesWant{noOutput: []string{"Agent:"}, noFiles: []string{"AGENTS.md"}},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			repoDir := t.TempDir()
-			restore := overrideNewRunner(repoRootRunner(repoDir, func(args ...string) (string, error) {
-				if args[0] == cmdSymbolicRef {
-					return refsRemotesOriginMain, nil
-				}
-				return "", nil
-			}))
-			defer restore()
-
-			if tt.preExisting {
-				if err := os.MkdirAll(filepath.Join(repoDir, config.DirName), 0755); err != nil {
-					t.Fatal(err)
-				}
-			}
-			if tt.runTwice {
-				cmd1, _ := newTestCmd()
-				cmd1.Flags().Bool(flagAgents, true, "")
-				if err := initCmd.RunE(cmd1, nil); err != nil {
-					t.Fatalf("first initCmd.RunE: %v", err)
-				}
-			}
-
-			cmd, buf := newTestCmd()
-			if tt.withFlag {
-				cmd.Flags().Bool(flagAgents, true, "")
-			}
-			if err := initCmd.RunE(cmd, nil); err != nil {
-				t.Fatalf("initCmd.RunE: %v", err)
-			}
-			assertAgentFilesResult(t, repoDir, buf.String(), tt.want)
-		})
-	}
-}
-
-func TestInitAgentsErrors(t *testing.T) {
-	tests := []struct {
-		name        string
-		blockPath   string // relative path to pre-create as a directory to trigger the error
-		wantErrFrag string // substring expected in the error
-	}{
-		{
-			name:        "install error (AGENTS.md is a directory)",
-			blockPath:   "AGENTS.md",
-			wantErrFrag: "agent files",
-		},
-		{
-			// Pre-create .mcp.json as a directory so os.ReadFile returns EISDIR (not ErrNotExist),
-			// causing RegisterMCPProject to fail after InstallProject succeeds.
-			name:        "MCP register error (.mcp.json is a directory)",
-			blockPath:   ".mcp.json",
-			wantErrFrag: "mcp servers",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			repoDir := t.TempDir()
-			if err := os.MkdirAll(filepath.Join(repoDir, tt.blockPath), 0750); err != nil {
-				t.Fatal(err)
-			}
-
-			repoName := filepath.Base(repoDir)
-			wtPath := filepath.Join(filepath.Dir(repoDir), repoName+"-worktrees")
-			t.Cleanup(func() { os.RemoveAll(wtPath) })
-
-			r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-				if args[0] == cmdSymbolicRef {
-					return refsRemotesOriginMain, nil
-				}
-				return "", nil
-			})
-			restore := overrideNewRunner(r)
-			defer restore()
-
-			cmd, _ := newTestCmd()
-			cmd.Flags().Bool(flagAgents, true, "")
-			err := initCmd.RunE(cmd, nil)
-			if err == nil {
-				t.Fatalf("expected error containing %q, got nil", tt.wantErrFrag)
-			}
-			if !strings.Contains(err.Error(), tt.wantErrFrag) {
-				t.Errorf("error = %q, want %q", err.Error(), tt.wantErrFrag)
-			}
-		})
 	}
 }
 
@@ -536,584 +366,5 @@ func TestInitFreshGitignoreAlreadyPresent(t *testing.T) {
 	out := buf.String()
 	if !strings.Contains(out, "already in .gitignore") {
 		t.Errorf("output should say 'already in .gitignore', got:\n%s", out)
-	}
-}
-
-// --- Tests for new 3-tier flag surface ---
-
-func TestInitGlobalInstall(t *testing.T) {
-	home, restore := setupGlobalInit(t)
-	defer restore()
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagGlobal, true, "")
-
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "user") {
-		t.Errorf("output should mention 'user', got:\n%s", out)
-	}
-
-	// Verify Claude skill was created
-	claudePath := filepath.Join(home, ".claude", "skills", "rimba", "SKILL.md")
-	if _, err := os.Stat(claudePath); os.IsNotExist(err) {
-		t.Errorf("global claude skill not created at %s", claudePath)
-	}
-}
-
-func TestInitGlobalNoRepoRequired(t *testing.T) {
-	_, restore := setupGlobalInit(t)
-	defer restore()
-
-	cmd, _ := newTestCmd()
-	cmd.Flags().Bool(flagGlobal, true, "")
-
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("-g should succeed outside a git repo, got: %v", err)
-	}
-}
-
-func TestInitGlobalUninstall(t *testing.T) {
-	home, restore := setupGlobalInit(t)
-	defer restore()
-
-	cmd1, _ := newTestCmd()
-	cmd1.Flags().Bool(flagGlobal, true, "")
-	if err := initCmd.RunE(cmd1, nil); err != nil {
-		t.Fatalf("install: %v", err)
-	}
-
-	// Uninstall
-	cmd2, buf := newTestCmd()
-	cmd2.Flags().Bool(flagGlobal, true, "")
-	cmd2.Flags().Bool(flagUninstall, true, "")
-	if err := initCmd.RunE(cmd2, nil); err != nil {
-		t.Fatalf("uninstall: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "Removed") {
-		t.Errorf("output should say 'Removed', got:\n%s", out)
-	}
-
-	// Claude skill should be gone
-	claudePath := filepath.Join(home, ".claude", "skills", "rimba", "SKILL.md")
-	if _, err := os.Stat(claudePath); err == nil {
-		t.Error("claude skill should be removed after uninstall")
-	}
-}
-
-func TestInitAgentsLocalAddsGitignore(t *testing.T) {
-	repoDir := t.TempDir()
-
-	r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-		if args[0] == cmdSymbolicRef {
-			return refsRemotesOriginMain, nil
-		}
-		return "", nil
-	})
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	cmd, _ := newTestCmd()
-	cmd.Flags().Bool(flagAgents, true, "")
-	cmd.Flags().Bool(flagLocal, true, "")
-
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	data, err := os.ReadFile(filepath.Join(repoDir, ".gitignore"))
-	if err != nil {
-		t.Fatalf("read .gitignore: %v", err)
-	}
-	gitignore := string(data)
-	// AGENTS.md should be in .gitignore (local tier)
-	if !strings.Contains(gitignore, "AGENTS.md") {
-		t.Errorf(".gitignore should contain AGENTS.md, got:\n%s", gitignore)
-	}
-}
-
-func TestInitAgentsUninstall(t *testing.T) {
-	repoDir := t.TempDir()
-
-	r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-		if args[0] == cmdSymbolicRef {
-			return refsRemotesOriginMain, nil
-		}
-		return "", nil
-	})
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	// Install
-	cmd1, _ := newTestCmd()
-	cmd1.Flags().Bool(flagAgents, true, "")
-	if err := initCmd.RunE(cmd1, nil); err != nil {
-		t.Fatalf("install: %v", err)
-	}
-	agentsPath := filepath.Join(repoDir, "AGENTS.md")
-	if _, err := os.Stat(agentsPath); os.IsNotExist(err) {
-		t.Fatal("AGENTS.md should exist after install")
-	}
-
-	// Uninstall
-	cmd2, buf := newTestCmd()
-	cmd2.Flags().Bool(flagAgents, true, "")
-	cmd2.Flags().Bool(flagUninstall, true, "")
-	if err := initCmd.RunE(cmd2, nil); err != nil {
-		t.Fatalf("uninstall: %v", err)
-	}
-	if !strings.Contains(buf.String(), "Removed") {
-		t.Errorf("output should say 'Removed', got:\n%s", buf.String())
-	}
-	if _, err := os.Stat(agentsPath); err == nil {
-		t.Error("AGENTS.md should be removed after uninstall")
-	}
-}
-
-func TestInitAgentsLocalUninstall(t *testing.T) {
-	repoDir := t.TempDir()
-
-	r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-		if args[0] == cmdSymbolicRef {
-			return refsRemotesOriginMain, nil
-		}
-		return "", nil
-	})
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	// Install local
-	cmd1, _ := newTestCmd()
-	cmd1.Flags().Bool(flagAgents, true, "")
-	cmd1.Flags().Bool(flagLocal, true, "")
-	if err := initCmd.RunE(cmd1, nil); err != nil {
-		t.Fatalf("install: %v", err)
-	}
-	agentsPath := filepath.Join(repoDir, "AGENTS.md")
-	if _, err := os.Stat(agentsPath); os.IsNotExist(err) {
-		t.Fatal("AGENTS.md should exist after local install")
-	}
-
-	// Uninstall local
-	cmd2, buf := newTestCmd()
-	cmd2.Flags().Bool(flagAgents, true, "")
-	cmd2.Flags().Bool(flagLocal, true, "")
-	cmd2.Flags().Bool(flagUninstall, true, "")
-	if err := initCmd.RunE(cmd2, nil); err != nil {
-		t.Fatalf("uninstall: %v", err)
-	}
-	if !strings.Contains(buf.String(), "Removed") {
-		t.Errorf("output should say 'Removed', got:\n%s", buf.String())
-	}
-	if _, err := os.Stat(agentsPath); err == nil {
-		t.Error("AGENTS.md should be removed after local uninstall")
-	}
-}
-
-func TestInitFlagValidationErrors(t *testing.T) {
-	repoDir := t.TempDir()
-
-	r := repoRootRunner(repoDir, func(_ ...string) (string, error) { return "", nil })
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	tests := []struct {
-		name    string
-		setup   func(*cobra.Command)
-		wantErr string
-	}{
-		{
-			name: "--local without --agents",
-			setup: func(cmd *cobra.Command) {
-				cmd.Flags().Bool(flagLocal, true, "")
-			},
-			wantErr: "--local requires --agents",
-		},
-		{
-			name: "--uninstall without target",
-			setup: func(cmd *cobra.Command) {
-				cmd.Flags().Bool(flagUninstall, true, "")
-			},
-			wantErr: "--uninstall requires",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			cmd, _ := newTestCmd()
-			tt.setup(cmd)
-			err := initCmd.RunE(cmd, nil)
-			if err == nil {
-				t.Fatal("expected error, got nil")
-			}
-			if !strings.Contains(err.Error(), tt.wantErr) {
-				t.Errorf("error = %q, want to contain %q", err.Error(), tt.wantErr)
-			}
-		})
-	}
-}
-
-// --- MCP registration tests ---
-
-func TestInitGlobalRegistersMCP(t *testing.T) {
-	home, restore := setupGlobalInit(t)
-	defer restore()
-
-	// Pre-seed .claude/settings.json so MCP registration doesn't skip it
-	claudeDir := filepath.Join(home, ".claude")
-	if err := os.MkdirAll(claudeDir, 0750); err != nil {
-		t.Fatalf("mkdir: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(claudeDir, "settings.json"), []byte("{}\n"), 0600); err != nil {
-		t.Fatalf("write settings.json: %v", err)
-	}
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagGlobal, true, "")
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "Agent files:") {
-		t.Errorf("output should contain 'Agent files:', got:\n%s", out)
-	}
-	if !strings.Contains(out, "MCP server:") {
-		t.Errorf("output should contain 'MCP server:', got:\n%s", out)
-	}
-	if !strings.Contains(out, filepath.Join(".claude", "settings.json")) {
-		t.Errorf("output should mention settings.json, got:\n%s", out)
-	}
-	if !strings.Contains(out, "registered") {
-		t.Errorf("output should say 'registered', got:\n%s", out)
-	}
-
-	// Verify the file was patched
-	data, err := os.ReadFile(filepath.Join(claudeDir, "settings.json"))
-	if err != nil {
-		t.Fatalf("read settings.json: %v", err)
-	}
-	var cfg map[string]any
-	if err := json.Unmarshal(data, &cfg); err != nil {
-		t.Fatalf("parse settings.json: %v", err)
-	}
-	servers, ok := cfg["mcpServers"].(map[string]any)
-	if !ok {
-		t.Fatal("mcpServers missing in settings.json")
-	}
-	entry, ok := servers["rimba"].(map[string]any)
-	if !ok {
-		t.Fatal("mcpServers.rimba missing")
-	}
-	if entry["command"] != "rimba" {
-		t.Errorf("command = %v, want rimba", entry["command"])
-	}
-}
-
-func TestInitGlobalMCPSkippedWhenNoConfig(t *testing.T) {
-	_, restore := setupGlobalInit(t)
-	defer restore()
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagGlobal, true, "")
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "skipped") {
-		t.Errorf("output should contain 'skipped' for absent config files, got:\n%s", out)
-	}
-}
-
-func TestInitGlobalUninstallRemovesMCP(t *testing.T) {
-	home, restore := setupGlobalInit(t)
-	defer restore()
-
-	claudeDir := filepath.Join(home, ".claude")
-	if err := os.MkdirAll(claudeDir, 0750); err != nil {
-		t.Fatalf("mkdir: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(claudeDir, "settings.json"), []byte("{}\n"), 0600); err != nil {
-		t.Fatalf("write settings.json: %v", err)
-	}
-
-	// Install
-	cmd1, _ := newTestCmd()
-	cmd1.Flags().Bool(flagGlobal, true, "")
-	if err := initCmd.RunE(cmd1, nil); err != nil {
-		t.Fatalf("install: %v", err)
-	}
-
-	// Uninstall
-	cmd2, buf := newTestCmd()
-	cmd2.Flags().Bool(flagGlobal, true, "")
-	cmd2.Flags().Bool(flagUninstall, true, "")
-	if err := initCmd.RunE(cmd2, nil); err != nil {
-		t.Fatalf("uninstall: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "unregistered") {
-		t.Errorf("output should say 'unregistered', got:\n%s", out)
-	}
-
-	// rimba key should be gone
-	data, err := os.ReadFile(filepath.Join(claudeDir, "settings.json"))
-	if err != nil {
-		t.Fatalf("read: %v", err)
-	}
-	var cfg map[string]any
-	if err := json.Unmarshal(data, &cfg); err != nil {
-		t.Fatalf("parse: %v", err)
-	}
-	if servers, ok := cfg["mcpServers"].(map[string]any); ok {
-		if _, found := servers["rimba"]; found {
-			t.Error("rimba key should be removed after uninstall")
-		}
-	}
-}
-
-func TestInitAgentsMCPRegistration(t *testing.T) {
-	repoDir := t.TempDir()
-
-	// Pre-seed .mcp.json so registration doesn't skip
-	if err := os.WriteFile(filepath.Join(repoDir, ".mcp.json"), []byte("{}\n"), 0600); err != nil {
-		t.Fatalf("write .mcp.json: %v", err)
-	}
-
-	r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-		if args[0] == cmdSymbolicRef {
-			return refsRemotesOriginMain, nil
-		}
-		return "", nil
-	})
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagAgents, true, "")
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "MCP server:") {
-		t.Errorf("output should contain 'MCP server:', got:\n%s", out)
-	}
-	if !strings.Contains(out, ".mcp.json") {
-		t.Errorf("output should mention .mcp.json, got:\n%s", out)
-	}
-
-	data, err := os.ReadFile(filepath.Join(repoDir, ".mcp.json"))
-	if err != nil {
-		t.Fatalf("read .mcp.json: %v", err)
-	}
-	var cfg map[string]any
-	if err := json.Unmarshal(data, &cfg); err != nil {
-		t.Fatalf("parse .mcp.json: %v", err)
-	}
-	servers, ok := cfg["mcpServers"].(map[string]any)
-	if !ok {
-		t.Fatal("mcpServers missing in .mcp.json")
-	}
-	if _, ok := servers["rimba"]; !ok {
-		t.Error("mcpServers.rimba should be registered in .mcp.json")
-	}
-}
-
-func TestInitAgentsLocalSkipsMCP(t *testing.T) {
-	repoDir := t.TempDir()
-
-	r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-		if args[0] == cmdSymbolicRef {
-			return refsRemotesOriginMain, nil
-		}
-		return "", nil
-	})
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagAgents, true, "")
-	cmd.Flags().Bool(flagLocal, true, "")
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if strings.Contains(out, "MCP server:") {
-		t.Errorf("--agents --local should not output 'MCP server:' section, got:\n%s", out)
-	}
-}
-
-func TestInitGlobalMCPErrorPropagated(t *testing.T) {
-	home, restore := setupGlobalInit(t)
-	defer restore()
-
-	// Write malformed JSON to settings.json — triggers patchJSON error
-	claudeDir := filepath.Join(home, ".claude")
-	if err := os.MkdirAll(claudeDir, 0750); err != nil {
-		t.Fatalf("mkdir: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(claudeDir, "settings.json"), []byte("not json"), 0600); err != nil {
-		t.Fatalf("write: %v", err)
-	}
-
-	cmd, _ := newTestCmd()
-	cmd.Flags().Bool(flagGlobal, true, "")
-	err := initCmd.RunE(cmd, nil)
-	if err == nil {
-		t.Fatal("expected error for malformed MCP config, got nil")
-	}
-	if !strings.Contains(err.Error(), "mcp servers") {
-		t.Errorf("error should mention 'mcp servers', got: %v", err)
-	}
-}
-
-func TestInitPersonalFreshInit(t *testing.T) {
-	repoDir := t.TempDir()
-
-	r := repoRootRunner(repoDir, func(args ...string) (string, error) {
-		if args[0] == cmdSymbolicRef {
-			return refsRemotesOriginMain, nil
-		}
-		return "", nil
-	})
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagPersonal, true, "")
-
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "Initialized rimba") {
-		t.Errorf("output = %q, want 'Initialized rimba'", out)
-	}
-
-	// Verify .rimba/settings.toml was created
-	teamPath := filepath.Join(repoDir, config.DirName, config.TeamFile)
-	if _, err := os.Stat(teamPath); os.IsNotExist(err) {
-		t.Errorf("settings.toml not created at %s", teamPath)
-	}
-
-	// Verify .rimba/settings.local.toml was NOT created
-	localPath := filepath.Join(repoDir, config.DirName, config.LocalFile)
-	if _, err := os.Stat(localPath); err == nil {
-		t.Error("settings.local.toml should not be created in personal mode")
-	}
-
-	// Verify .gitignore has .rimba/ not .rimba/settings.local.toml
-	data, err := os.ReadFile(filepath.Join(repoDir, ".gitignore"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	content := string(data)
-	dirEntry := config.DirName + "/"
-	if !strings.Contains(content, dirEntry) {
-		t.Errorf(".gitignore should contain %q, got:\n%s", dirEntry, content)
-	}
-	localEntry := filepath.Join(config.DirName, config.LocalFile)
-	if strings.Contains(content, localEntry) {
-		t.Errorf(".gitignore should not contain %q in personal mode, got:\n%s", localEntry, content)
-	}
-}
-
-func TestInitPersonalMigration(t *testing.T) {
-	repoDir := t.TempDir()
-
-	// Pre-create legacy .rimba.toml
-	legacyCfg := &config.Config{WorktreeDir: "../wt", DefaultSource: "main"}
-	if err := config.Save(filepath.Join(repoDir, config.FileName), legacyCfg); err != nil {
-		t.Fatalf("save legacy config: %v", err)
-	}
-
-	// Pre-create .gitignore with legacy entry
-	if err := os.WriteFile(filepath.Join(repoDir, ".gitignore"), []byte(config.FileName+"\n"), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	r := repoRootRunner(repoDir, func(_ ...string) (string, error) { return "", nil })
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	cmd, buf := newTestCmd()
-	cmd.Flags().Bool(flagPersonal, true, "")
-
-	if err := initCmd.RunE(cmd, nil); err != nil {
-		t.Fatalf("initCmd.RunE: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "Migrated rimba config") {
-		t.Errorf("output should contain 'Migrated rimba config', got:\n%s", out)
-	}
-
-	// Verify legacy file is gone
-	if _, err := os.Stat(filepath.Join(repoDir, config.FileName)); !os.IsNotExist(err) {
-		t.Error("legacy .rimba.toml should have been removed")
-	}
-
-	// Verify .rimba/settings.toml exists and is loadable
-	cfg, err := config.LoadDir(filepath.Join(repoDir, config.DirName))
-	if err != nil {
-		t.Fatalf("LoadDir after migration: %v", err)
-	}
-	if cfg.WorktreeDir != "../wt" {
-		t.Errorf("WorktreeDir = %q, want %q", cfg.WorktreeDir, "../wt")
-	}
-
-	// Verify .rimba/settings.local.toml was NOT created
-	localPath := filepath.Join(repoDir, config.DirName, config.LocalFile)
-	if _, err := os.Stat(localPath); err == nil {
-		t.Error("settings.local.toml should not be created in personal mode")
-	}
-
-	// Verify .gitignore updated with .rimba/ not legacy entry
-	data, err := os.ReadFile(filepath.Join(repoDir, ".gitignore"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	content := string(data)
-	if strings.Contains(content, config.FileName) {
-		t.Error(".gitignore should not contain legacy entry after migration")
-	}
-	dirEntry := config.DirName + "/"
-	if !strings.Contains(content, dirEntry) {
-		t.Errorf(".gitignore should contain %q, got:\n%s", dirEntry, content)
-	}
-	localEntry := filepath.Join(config.DirName, config.LocalFile)
-	if strings.Contains(content, localEntry) {
-		t.Errorf(".gitignore should not contain %q in personal mode, got:\n%s", localEntry, content)
-	}
-}
-
-func TestInitRepoRootError(t *testing.T) {
-	r := &mockRunner{
-		run: func(args ...string) (string, error) {
-			if len(args) >= 2 && args[1] == cmdShowToplevel {
-				return "", errors.New("not a git repository")
-			}
-			return "", nil
-		},
-		runInDir: noopRunInDir,
-	}
-	restore := overrideNewRunner(r)
-	defer restore()
-
-	cmd, _ := newTestCmd()
-
-	err := initCmd.RunE(cmd, nil)
-	if err == nil {
-		t.Fatal("expected error when repo root fails")
 	}
 }
